+++ src/State/Provider/ObjectMapperProvider.php
@@ -15,7 +15,7 @@ namespace ApiPlatform\State\Provider;
 
 use ApiPlatform\Metadata\Operation;
 use ApiPlatform\Metadata\Util\CloneTrait;
-use ApiPlatform\State\Pagination\ArrayPaginator;
+use ApiPlatform\State\Pagination\MappedObjectPaginator;
 use ApiPlatform\State\Pagination\PaginatorInterface;
 use ApiPlatform\State\ProviderInterface;
 use Symfony\Component\ObjectMapper\ObjectMapperInterface;
@@ -39,6 +39,7 @@ final class ObjectMapperProvider implements ProviderInterface
     public function provide(Operation $operation, array $uriVariables = [], array $context = []): object|array|null
     {
         $data = $this->decorated->provide($operation, $uriVariables, $context);
+        $class = $operation->getOutput()['class'] ?? $operation->getClass();
 
         if (!$this->objectMapper || !$operation->canMap()) {
             return $data;
@@ -52,15 +53,23 @@ final class ObjectMapperProvider implements ProviderInterface
         $request?->attributes->set('mapped_data', $data);
 
         if ($data instanceof PaginatorInterface) {
-            $data = new ArrayPaginator(array_map(fn ($v) => $this->objectMapper->map($v, $operation->getClass()), iterator_to_array($data)), 0, \count($data));
+            $data = new MappedObjectPaginator(
+                iterator_to_array($data),
+                $this->objectMapper,
+                $class,
+                $data->getTotalItems(),
+                $data->getCurrentPage(),
+                $data->getLastPage(),
+                $data->getItemsPerPage(),
+            );
         } elseif (\is_array($data)) {
             foreach ($data as &$v) {
                 if (\is_object($v)) {
-                    $v = $this->objectMapper->map($v, $operation->getClass());
+                    $v = $this->objectMapper->map($v, $class);
                 }
             }
         } else {
-            $data = $this->objectMapper->map($data, $operation->getClass());
+            $data = $this->objectMapper->map($data, $class);
         }
 
         $request?->attributes->set('data', $data);