+++ src/Symfony/Bundle/Resources/config/state/object_mapper.php
@@ -18,13 +18,7 @@ return function (ContainerConfigurator $container) {
 
     $services->set('api_platform.object_mapper.metadata_factory', 'Symfony\Component\ObjectMapper\Metadata\ReflectionObjectMapperMetadataFactory');
 
-    $services->set('api_platform.object_mapper', 'Symfony\Component\ObjectMapper\ObjectMapper')
-        ->args([
-            service('api_platform.object_mapper.metadata_factory'),
-            service('property_accessor')->nullOnInvalid(),
-            tagged_locator('object_mapper.transform_callable'),
-            tagged_locator('object_mapper.condition_callable'),
-        ]);
+    $services->alias('api_platform.object_mapper', 'object_mapper');
 
     $services->set('api_platform.object_mapper.relation', 'ApiPlatform\State\ObjectMapper\ObjectMapper')
         ->decorate('api_platform.object_mapper', null, -255)